feat: add Redis Stack module
